pyproject.toml
[project]
name = "myhikes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.85",
    "gpxpy>=1.6.2",
    "jinja2>=3.1.6",
    "piexif>=1.1.3",
    "pillow>=12.2.0",
    "pytest>=9.0.3",
]

[project.scripts]
hikes = "generator.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["generator"]
